various bits of code tidying
git-svn-id: https://jmoney.svn.sf.net/svnroot/jmoney/trunk@462 b97006a3-7e3d-0410-986d-ea81e883d316

+++ net.sf.jmoney/src/net/sf/jmoney/fields/AmountEditor.java
@@ -89,7 +89,7 @@ public class AmountEditor implements IPropertyControl {
     		// Some amounts may be of type Long, not long, so that 
     		// they can be null, so we must get the property
     		// value as a Long.
-    		Long amount = (Long) fObject.getPropertyValue(amountPropertyAccessor);
+    		Long amount = fObject.getPropertyValue(amountPropertyAccessor);
     		if (amount == null) {
     			propertyControl.setText("");
     		} else {
